+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## v0.0.2 (unreleased)
 
-Nothing has changed yet.
+* Added a ``remove_index`` to remove unwanted index,
 
 ## v0.0.1 (2014-10-26)